+++ src/api/java/blusunrize/immersiveengineering/api/Lib.java
@@ -75,6 +75,7 @@ public class Lib
 	public static final String GUIID_ToolboxBlock = "toolboxblock";
 	public static final String GUIID_RedstoneConnector = "redstoneconnector";
 	public static final String GUIID_RedstoneProbe = "redstoneprobe";
+	public static final String GUIID_RedstoneStateCell = "redstonestatecell";
 	//Items
 	public static final String GUIID_Revolver = "revolver";
 	public static final String GUIID_Toolbox = "toolbox";