+++ changelog.md
@@ -1,4 +1,5 @@
 ##### Version 1.20.4-11.5.1-pre
+- Add the redstone state cell for storing signals (BluSunrize)
 - Fix waterlogged posts not dropping items (BluSunrize)
 - Fix treated windows and feedthrough insulators letting rain through (BluSunrize, voidsong-dragonfly)
 - Fix crates not syncing their names to the client (BluSunrize)